pyproject.toml
# 构建工具
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

# 项目信息和依赖管理
[project]
name = "api-automated-testing" # 项目名
version = "1.0"      # 版本号
dependencies = [#依赖
    "requests>=2.32.4",
    "pytest>=8.3.5",
    "openpyxl>=3.1.3",
    "pytest-order>=1.3.0",
    "pytest-rerunfailures>=14.0",
    "allure-pytest>=2.15.0",
    "allure-python-commons>=2.15.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1"
]